+++ main-service/src/main/java/ru/practicum/ewm/events/utils/ValidatorDefaultFields.java
@@ -30,5 +30,4 @@ public class ValidatorDefaultFields {
             }
         }
     }
-
 }
